Fixed index-fields in migration-files

+++ resources/migrations/simple/2019_04_16_065614_create_cart_sessions_table.php
@@ -16,7 +16,7 @@ class CreateCartSessionsTable extends Migration
         Schema::create('ls_cart_sessions', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('session_token')->unique();
-            $table->unsignedInteger('shipping_option_id');
+            $table->unsignedBigInteger('shipping_option_id');
             $table->foreign('shipping_option_id')->references('id')->on('ls_shipping_options')->onDelete('restrict');
             $table->json('billing_address')->nullable();
             $table->json('shipping_address')->nullable();
