+++ resources/migrations/simple/2019_04_16_065648_create_vouchers_table.php
@@ -19,7 +19,7 @@ class CreateVouchersTable extends Migration
             $table->string('code')->unique();
             $table->unsignedDecimal('discount_amount', 8 ,2)->nullable();
             $table->unsignedDecimal('discount_percentage', 8 ,2)->nullable();
-            $table->unsignedInteger('qty')->nullable();
+            $table->unsignedBigInteger('qty')->nullable();
             $table->timestamp('expires_on')->nullable();
             $table->boolean('active')->default(0);
