+++ resources/migrations/simple/2019_04_16_065933_create_product_sale_table.php
@@ -15,9 +15,9 @@ class CreateProductSaleTable extends Migration
     {
         Schema::create('ls_product_sale', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('ls_products')->onDelete('cascade');
-            $table->unsignedInteger('sale_id');
+            $table->unsignedBigInteger('sale_id');
             $table->foreign('sale_id')->references('id')->on('ls_sales')->onDelete('cascade');
         });
     }
